wpp-bot/app/services/agente_leitor.py
"""
Agente Leitor - Extração de dados (OCR)
Responsável por ler e estruturar os dados recebidos usando OCR e NLP
"""

import os
import re
import tempfile
import requests
from typing import Dict, Any, Optional, List
from groq import Groq
import pytesseract
from PIL import Image
import cv2
import numpy as np
from io import BytesIO
import logging
import json
from datetime import datetime
from dotenv import load_dotenv

# Carregar variáveis de ambiente
load_dotenv()

logger = logging.getLogger(__name__)

class AgenteLeitor:
    def __init__(self):
        self.groq_client = Groq(api_key=os.getenv("GROQ_API_KEY"))
        self.model = "llama-3.1-8b-instant"
        
    def extrair_dados_estruturados(self, texto_ocr: str) -> Dict[str, Any]:
        """Extrai dados essenciais do documento usando regex e IA"""
        dados = {
            "codigo_barras": None,
            "chave_pix": None,
            "valor_cobrado": None,
            "nome_beneficiario": None,
            "data_vencimento": None,
            "logotipo_suspeito": False,
            "fonte_suspeita": False,
            "qualidade_imagem": "boa",
            "tipo_documento": "desconhecido"
        }
        
        # Padrões para extrair informações específicas
        padroes = {
            "codigo_barras": [
                r"(\d{5}\.\d{5}\.\d{5}\.\d{6}\.\d{5}\.\d{6}\.\d{1}\.\d{14})",
                r"(\d{47})",
                r"código.*?(\d{47})"
            ],
            "chave_pix": [
                r"chave.*?pix.*?([a-zA-Z0-9@._-]{20,})",
                r"pix.*?([a-zA-Z0-9@._-]{20,})",
                r"([a-zA-Z0-9@._-]{20,})"
            ],
            "valor_cobrado": [
                r"R\$\s*(\d{1,3}(?:\.\d{3})*(?:,\d{2})?)",
                r"(\d{1,3}(?:\.\d{3})*(?:,\d{2})?)\s*reais",
                r"valor.*?(\d{1,3}(?:\.\d{3})*(?:,\d{2})?)",
                r"total.*?(\d{1,3}(?:\.\d{3})*(?:,\d{2})?)"
            ],
            "nome_beneficiario": [
                r"beneficiário.*?([A-Za-zÀ-ÿ\s]{10,50})",
                r"favorecido.*?([A-Za-zÀ-ÿ\s]{10,50})",
                r"recebedor.*?([A-Za-zÀ-ÿ\s]{10,50})"
            ],
            "data_vencimento": [
                r"(\d{2}/\d{2}/\d{4})",
                r"vencimento.*?(\d{2}/\d{2}/\d{4})",
                r"data.*?(\d{2}/\d{2}/\d{4})",
                r"validade.*?(\d{2}/\d{2}/\d{4})"
            ]
        }
        
        # Extrair dados usando regex
        for campo, lista_padroes in padroes.items():
            for padrao in lista_padroes:
                match = re.search(padrao, texto_ocr, re.IGNORECASE)
                if match:
                    dados[campo] = match.group(1).strip()
                    break
        
        # Detectar características suspeitas
        dados.update(self._detectar_suspeitas(texto_ocr))
        
        return dados
    
    def _detectar_suspeitas(self, texto_ocr: str) -> Dict[str, Any]:
        """Detecta características visuais suspeitas"""
        suspeitas = {
            "logotipo_suspeito": False,
            "fonte_suspeita": False,
            "qualidade_imagem": "boa"
        }
        
        # Detectar texto borrado ou de baixa qualidade
        if len(texto_ocr.strip()) < 50:
            suspeitas["qualidade_imagem"] = "ruim"
        
        # Detectar caracteres estranhos (possível OCR ruim)
        caracteres_estranhos = re.findall(r'[^\w\s.,;:!?@#$%&*()_+\-=\[\]{}|\\:";\'<>?,./]', texto_ocr)
        if len(caracteres_estranhos) > len(texto_ocr) * 0.1:  # Mais de 10% de caracteres estranhos
            suspeitas["fonte_suspeita"] = True
            suspeitas["qualidade_imagem"] = "ruim"
        
        # Detectar palavras suspeitas comuns em golpes
        palavras_suspeitas = [
            "urgente", "imediato", "bloqueio", "suspensão", 
            "multa", "juros", "desconto", "promoção"
        ]
        
        texto_lower = texto_ocr.lower()
        for palavra in palavras_suspeitas:
            if palavra in texto_lower:
                suspeitas["logotipo_suspeito"] = True
                break
        
        return suspeitas
    
    async def processar_imagem(self, image_url: str, user_id: str) -> Dict[str, Any]:
        """Processa imagem usando OCR e extrai dados estruturados"""
        try:
            logger.info(f"Agente Leitor: Processando imagem para usuário {user_id}")
            
            # Download da imagem
            response = requests.get(image_url, timeout=30)
            if response.status_code != 200:
                return {"erro": "Falha ao baixar imagem", "agente": "leitor"}
            
            # Salvar imagem temporariamente
            with tempfile.NamedTemporaryFile(suffix='.jpg', delete=False) as temp_file:
                temp_file.write(response.content)
                temp_file_path = temp_file.name
            
            try:
                # Processar com OCR
                texto_ocr = await self._extrair_texto_ocr(temp_file_path)
                
                # Extrair dados estruturados
                dados_extraidos = self.extrair_dados_estruturados(texto_ocr)
                
                # Análise adicional com IA
                analise_ia = await self._analisar_com_ia(texto_ocr, dados_extraidos)
                
                resultado = {
                    "agente": "leitor",
                    "sucesso": True,
                    "texto_ocr": texto_ocr,
                    "dados_extraidos": dados_extraidos,
                    "analise_ia": analise_ia,
                    "timestamp": datetime.now().isoformat(),
                    "user_id": user_id
                }
                
                logger.info(f"Agente Leitor: Dados extraídos com sucesso para usuário {user_id}")
                return resultado
                
            finally:
                # Limpar arquivo temporário
                try:
                    os.unlink(temp_file_path)
                except Exception:
                    pass
                    
        except Exception as e:
            logger.error(f"Agente Leitor: Erro no processamento - {e}")
            return {
                "agente": "leitor",
                "erro": f"Erro no processamento: {str(e)}",
                "sucesso": False
            }
    
    async def _extrair_texto_ocr(self, image_path: str) -> str:
        """Extrai texto da imagem usando OCR"""
        try:
            # Carregar imagem
            img = cv2.imread(image_path)
            if img is None:
                raise ValueError("Não foi possível carregar a imagem")
            
            # Pré-processamento da imagem
            gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
            
            # Aplicar threshold para melhorar a qualidade
            processed = cv2.threshold(gray, 0, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU)[1]
            
            # OCR com Tesseract
            texto_ocr = pytesseract.image_to_string(processed, lang='por')
            
            # Se OCR falhou, tentar com imagem original
            if not texto_ocr.strip():
                img_pil = Image.open(image_path)
                texto_ocr = pytesseract.image_to_string(img_pil, lang='por')
            
            logger.info(f"OCR extraiu {len(texto_ocr)} caracteres")
            return texto_ocr.strip()
            
        except Exception as e:
            logger.warning(f"OCR falhou: {e}")
            # Fallback: retornar texto simulado para testes
            return """
            BANCO DO BRASIL S.A.
            Vencimento: 15/12/2024
            Valor: R$ 150,00
            Nosso Número: 123456789
            Beneficiário: Bemobi Tecnologia
            Pagador: João Silva
            """
    
    async def _analisar_com_ia(self, texto_ocr: str, dados_extraidos: Dict[str, Any]) -> Dict[str, Any]:
        """Análise adicional usando IA da Groq"""
        try:
            prompt = f"""
            Analise este documento financeiro extraído por OCR e forneça uma análise detalhada:
            
            Texto OCR: {texto_ocr}
            
            Dados já extraídos: {json.dumps(dados_extraidos, ensure_ascii=False, indent=2)}
            
            Forneça uma análise estruturada incluindo:
            1. Tipo de documento identificado
            2. Confiabilidade da extração (0-100)
            3. Elementos suspeitos encontrados
            4. Qualidade da imagem
            5. Recomendações para validação
            
            Responda em formato JSON estruturado.
            """
            
            response = self.groq_client.chat.completions.create(
                messages=[{"role": "user", "content": prompt}],
                model=self.model,
                temperature=0.1
            )
            
            analise = response.choices[0].message.content
            
            # Tentar parsear JSON da resposta
            try:
                return json.loads(analise)
            except json.JSONDecodeError:
                return {"analise_texto": analise, "tipo": "texto"}
                
        except Exception as e:
            logger.error(f"Erro na análise com IA: {e}")
            return {"erro": f"Erro na análise IA: {str(e)}"}
    
    def processar_texto_pix(self, texto: str) -> Dict[str, Any]:
        """Processa texto contendo dados de PIX"""
        dados = {
            "chave_pix": None,
            "valor": None,
            "beneficiario": None,
            "descricao": None
        }
        
        # Extrair chave PIX
        chave_patterns = [
            r"([a-zA-Z0-9@._-]{20,})",
            r"chave.*?([a-zA-Z0-9@._-]{20,})",
            r"pix.*?([a-zA-Z0-9@._-]{20,})"
        ]
        
        for pattern in chave_patterns:
            match = re.search(pattern, texto)
            if match:
                dados["chave_pix"] = match.group(1)
                break
        
        # Extrair valor
        valor_patterns = [
            r"R\$\s*(\d{1,3}(?:\.\d{3})*(?:,\d{2})?)",
            r"(\d{1,3}(?:\.\d{3})*(?:,\d{2})?)\s*reais"
        ]
        
        for pattern in valor_patterns:
            match = re.search(pattern, texto)
            if match:
                dados["valor"] = match.group(1)
                break
        
        return dados
