wpp-bot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
alembic==1.12.1
httpx==0.25.1
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
tenacity==8.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.2.1
Jinja2==3.1.2
pytz==2023.3.post1
loguru==0.7.2
groq==0.4.1
pytesseract==0.3.10
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
requests==2.31.0